Initial skeleton version of project (non-working)
This is an initial skeleton version of a non-working project, it builds
and there is an outline of what the library will become. It is meant to
be an allocator written in portable C that can allocate with various
algorithms within a pre-defined arena, that also provides more
functionality than the standard "malloc", "calloc", "realloc" and "free"
that are available in the C standard library.